Added functions and files.
Added a new utils.c where i moved the ft_exit and ft_atoi function. Added a new cleanup and get_time function. get_time is to get the time in miliseconds and cleanup is as it saysfor cleanup. Added a right int main and a hoin threads and create threads function in the philo.c. Moved the ft_exit function also to utils and in input.c(now init.c) i added the initilize functions for the mutex, philosophers and the initialize input