philosopher_routine added
Added the function philosopher_routine. If 1 philo he takes the fork and waits for his death, if odd numbered philos we wait 1ms so they dont take their forks simultaneously. Then in the while loop eating we call the philo eat function and check if finished after that with the check if finished function. Then the philo enters sleeping and then thinking and waiting for his turn to eat